feat: 타입 수

+++ src/main/java/com/example/doktoribackend/reading/domain/ReadingGenre.java
@@ -27,7 +27,7 @@ public class ReadingGenre extends BaseTimeEntity {
     private String name;
 
     @Column(name= "priority", nullable = false, columnDefinition = "TINYINT")
-    private Byte priority;
+    private Integer priority;
 
     @Column(name = "deleted_at")
     private LocalDateTime deletedAt;
